src/app/Bank_MS/home/home.component.ts
import { Component, OnInit } from '@angular/core';

import { Router } from '@angular/router';
import { AuthService } from 'src/app/services/auth.service';


@Component({
  selector: 'app-home',
  templateUrl: './home.component.html',
  styleUrls: ['./home.component.css']
})
export class HomeComponent implements OnInit {

  cid;

  constructor(
    private authService: AuthService,
    private router: Router
  ) { }

  ngOnInit(): void {
    this.cid = localStorage.getItem('CID');
  }

  logout() {
    this.authService.logout();
    this.router.navigateByUrl('/login');
  }

}
